+++ src/pages/payment/PaymentOffline.jsx
@@ -1,6 +1,6 @@
 // eslint-disable-next-line no-unused-vars
 import React from "react";
-import "./index.css";
+import "./PaymentOffline.css";
 
 function Payment() {
   return (
